+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog
 v1.4.3
 ------
 - fixing bug in parsing of MGI Genetic Marker file
+- fixing bug in parsing of orphanet genes file, en_product6.xml
 
 ------
 v1.4.2